refactor: BoardController 책임 분리 — BoardLikeController, DeletedBoardController 추출 (SRP Phase 2)
- BoardLikeController 신규: 좋아요 CRUD + 조회수/좋아요수 통계 (5개 엔드포인트)
- DeletedBoardController 신규: 삭제 관리 — 목록/복원/영구삭제 (3개 엔드포인트)
- BoardController: 20개 → 8개 엔드포인트로 축소, BoardLikeService 의존성 제거
- PrincipalUtil 추출: BoardController·CommentController 중복 getMemberId() 통합
